team_submissions/xinyi-li/requirements.txt
numpy
pandas
matplotlib
yfinance
ta
python-dotenv
google-generativeai